requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.1.0
Flask-Migrate>=4.0.0
Flask-Limiter>=3.5.0
Werkzeug>=2.3.0
Pillow>=10.0.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
email-validator>=2.0.0
bcrypt>=4.0.0
APScheduler>=3.10.0
gunicorn>=21.2.0